requirements.txt
tensorflow>=2.5.0
numpy>=1.19.2
flask>=2.0.0
flask-cors>=3.0.10
matplotlib>=3.3.4
scikit-learn>=0.24.1
pillow>=8.2.0
